Add mapper README. Improve code.
