+++ hardware/launch/rplidar_launch.launch
@@ -1,5 +1,5 @@
 <launch>
     <node name="lidar" pkg="rplidar" type="rplidarNode" output="screen">
-
+        <param name="frame_id" type="string" value="laser">
     </node>
 </launch>
